pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "This is a bot primarily made for learning discord.py and more...."
authors = [
    {name = "Anirban Nath"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py (>=2.5.0,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "wavelink (>=3.4.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "bot.main:run"